moved appointment hahandling functions and firestore resources to their own files
