+++ src/index.js
@@ -4,24 +4,6 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 
-import { initializeApp } from "firebase/app";
-import { getFirestore } from "firebase/firestore";
-// TODO: Add SDKs for Firebase products that you want to use
-// https://firebase.google.com/docs/web/setup#available-libraries
-
-// Your web app's Firebase configuration
-const firebaseConfig = {
-  apiKey: "AIzaSyCMOUBwIo7-2J1Q_GAwLJeNaO5ocixogGk",
-  authDomain: "curtiphotos-website.firebaseapp.com",
-  projectId: "curtiphotos-website",
-  storageBucket: "curtiphotos-website.firebasestorage.app",
-  messagingSenderId: "758029073029",
-  appId: "1:758029073029:web:69c5505c415cf5128ec7ec"
-};
-
-// Initialize Firebase
-const app = initializeApp(firebaseConfig);
-const db = getFirestore(app)
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
@@ -35,5 +17,3 @@ root.render(
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
-
-export { db, app };
